Add module ABI & API for `clear`ing tables (#4729)
# Description of Changes

Support is added to Rust, C#, C++, and TS modules.

# API and ABI breaking changes

None

# Expected complexity level and risk

1

# Testing

I wasn't able to regen module bindings for sdk tests, hitting an error
while doing so.

---------

+++ crates/core/src/db/relational_db.rs
@@ -1449,7 +1449,7 @@ impl RelationalDB {
     }
 
     /// Clears all rows from a table without dropping it.
-    pub fn clear_table(&self, tx: &mut MutTx, table_id: TableId) -> Result<usize, DBError> {
+    pub fn clear_table(&self, tx: &mut MutTx, table_id: TableId) -> Result<u64, DBError> {
         let rows_deleted = tx.clear_table(table_id)?;
         Ok(rows_deleted)
     }
